+++ src/network/handler/LoginHandler.ts
@@ -24,7 +24,7 @@ export default class LoginHandler {
         player.skin = packet.skin;
         player.device = packet.device;
 
-        player.sendPlayStatus(PlayStatus.LoginSuccess);
+        player.getPlayerConnection().sendPlayStatus(PlayStatus.LoginSuccess);
 
         const reason = server.getPermissionManager().isBanned(player);
         if (reason !== false) {
@@ -33,6 +33,6 @@ export default class LoginHandler {
         }
 
         let pk = new ResourcePacksInfoPacket();
-        player.sendDataPacket(pk);
+        player.getPlayerConnection().sendDataPacket(pk);
     }
 };
